WSI/tile_annotation_pipeline_repo/requirements.txt
numpy
pandas
tqdm
opencv-python
Pillow
matplotlib
pyyaml
scikit-image
openslide-python
